restrict age to only allow integers

+++ src/api.ts
@@ -9,7 +9,7 @@ const api = Router();
 
 const PayloadRequestSchema = z.object({
   name: z.string().min(1),
-  age: z.number().min(1),
+  age: z.number().int().min(1),
   timestamp: z.coerce.date()
 });
